# ProcesadorEnviosAPI/Models/OperadorLogistico.cs
using System;

namespace ProcesadorEnviosAPI.Models
{
    public class OperadorLogistico
    {
        public string Nombre { get; set; }

        public string Ciudad { get; set; }

        public string Provincia { get; set; }
    }
}